requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
tiktoken
openai
streamlit